backend/requirements.txt
langchain
openai
python-dotenv
langgraph
tiktoken
langchain-openai
flask
flask-cors
whisper
fastapi
uvicorn
python-multipart